typescriptification complete
